Fix unit tests in Android Studio
Sigh, it's always the some fucking story. "It's really Java", they said.
But then, there're always one or two missing classes like nio or in this
case class loader resources that can't be addressed if you happen to run
the tests from within Android Studio. This isn't even a Dalvik or ART
problem, but solely a fuckup in Gradle and the Android Studio
integration. Nonetheless, it leads to a death by a thousand paper cuts.
And I'm running out of bandaids.